requirements.txt
asgiref==3.5.0
boto3==1.21.10
botocore==1.24.10
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
Deprecated==1.2.13
Django==4.0.3
django-cors-headers==3.11.0
django-filter==21.1
django-oauth-toolkit==1.7.0
django-storages==1.12.3
djangorestframework==3.13.1
djongo==1.3.6
dnspython==2.2.0
drf-yasg==1.20.0
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
jmespath==0.10.0
jwcrypto==1.0
MarkupSafe==2.1.0
mongoengine==0.24.0
mysqlclient==2.1.0
oauthlib==3.2.0
packaging==21.3
Pillow==9.0.1
pycparser==2.21
pymongo==3.12.1
pyparsing==3.0.7
python-dateutil==2.8.2
python-environ==0.4.54
pytz==2021.3
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
s3transfer==0.5.2
six==1.16.0
sqlparse==0.2.4
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.8
whitenoise==6.0.0
wrapt==1.13.3
PyJWT==1.7.1
drf-jwt==1.19.0